server/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
google-generativeai==0.8.3
Werkzeug==3.0.1
flask-cors==4.0.0
python-magic==0.4.27
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.2
markdown==3.5.2
Pillow==10.2.0
plotly==5.18.0
base64io==1.0.3
protobuf==4.25.1